Installed react-router-dom, made Blackjack.js
