--- src/components/GameArea/GameArea.js
@@ -1,13 +1,0 @@
-import { useEffect, useState } from 'react'
-import './GameArea.scss'
-
-// Game Type, Win/Loss
-
-const GameArea = props => {
-    const [ gameType, setGameType ] = useState({ game: ''})
-    const [ winState, setWinState ] = useState({ win: false })
-
-    
-}
-
-export default GameArea
