--- src/components/PlayerArea/playerarea.scss
@@ -1,31 +1,0 @@
-#player-area-container {
-    margin: 100px 100px 50px 180px; 
-    clear: both; 
-    border: 1px solid #3b3b3b; 
-    height: 350px; 
-    width: 600px; 
-    border-radius: 2px; 
-
-}
-
-
-#Player1 {
-    text-align: left; 
-    font-size: 24px; 
-    margin: 20px;
-}
-
-
-
-
-
-
-
-#PlayerCard {
-    height: 200px; 
-    width: 300px; 
-    margin: 50px 10px 20px 20px; 
-    align-items: justify;
-    padding-right: 200px; 
-
-}
