--- src/components/navbar/NavBar.css
@@ -1,24 +1,0 @@
-.main-navbar {
-  padding: 1rem 0;
-  background: #3b3b3b;
-  color: #fff; }
-  .main-navbar a {
-    padding: 0.5rem;
-    color: inherit;
-    text-decoration: none;
-    border-radius: 2rem;
-    font-size: 1.6rem; }
-    .main-navbar a.current {
-      background: #555; }
-    .main-navbar a:hover {
-      color: inherit;
-      background: #999999;
-      text-decoration: none; }
-    .main-navbar a:active {
-      background: #dbdbdb;
-      color: #3b3b3b; }
-  .main-navbar .score-balance {
-    margin-left: auto; }
-    @media screen and (min-width: 992px) {
-      .main-navbar .score-balance .high-score {
-        border-right: 1px solid #fff; } }
